refactor: domain에 Reservation폴더 생성 및 추가해서 inport 전체적으로 수정됨
