+++ src/main/java/com/shinhanDS5gi/memento/domain/Reservation.java
@@ -1,7 +1,6 @@
-package com.shinhanDS5gi.memento.domain.Reservation;
+package com.shinhanDS5gi.memento.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.shinhanDS5gi.memento.domain.Mentos;
 import com.shinhanDS5gi.memento.domain.base.BaseStatus;
 import com.shinhanDS5gi.memento.domain.base.BaseTime;
 import com.shinhanDS5gi.memento.domain.member.Member;
